pyproject.toml
[tool.poetry]
name = "TLTester"
version = "0.1.0"
description = "This application controls Hioki's LCR meter (IM3536) and Sigma Kouki's stage controller (controller: shot-702, stage: SGSP-26-200) at the same time."
authors = ["5yutan5 <63651161+5yutan5@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/5yutan5/tensile-LCR-tester"
homepage = "https://github.com/5yutan5/tensile-LCR-tester"
packages = [{ include = "tester" }]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Environment :: X11 Applications :: Qt",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.scripts]
tltester = "tester.app.start:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
PySide6 = "^6.2.1"
psutil = "^5.8.0"
pyserial = "^3.5"
pyqtdarktheme = "^0.1.7"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
isort = "^5.10.0"
flake8 = "^4.0.1"
black = "^21.10b0"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 119
max-complexity = 10
extend-ignore = "E203"
per-file-ignores = ["**/__init__.py:F401"]

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119
